Trading/requirements.txt
python-dotenv>=1.0.0
PyYAML>=6.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.31.0
boto3>=1.28.0
pydantic>=2.0.0
pandas
numpy
scipy
yfinance
alpaca-py
tabulate
